+++ CHANGELOG.md
@@ -20,6 +20,7 @@ None
 * Fix issues with CTE inlining when using collection DML API
 * Only expose `Serializable` `EntityViewManager` wrapper in special methods or callbacks
 * Support `@PostCreate` and other lifecycle methods with default and even private visibility
+* Support `@ViewFilter` inheritance
 
 ### Backwards-incompatible changes